Implemented file_reader function
